% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ck_create_pert_params.R
\name{ck_create_pert_params}
\alias{ck_create_pert_params}
\title{ck_create_pert_params}
\usage{
ck_create_pert_params(big_n = NULL, small_n = NULL, ptab,
  stab = NULL, mtab = NULL)
}
\arguments{
\item{big_n}{(integer) large prime number used to derive cell keys from
record keys}

\item{small_n}{(integer) parameter that controlls how the columns of
\code{pTable} are scrolled}

\item{ptab}{a perturbation table generated with
\code{\link{ck_create_ptab}} or \code{\link{ck_create_free_ptable}}}

\item{stab}{table containing the noise component for numerical variables
generated by \code{\link{ck_generate_stab}}}

\item{mtab}{(numeric) vector containing parameters for the magnitude of
perturbation for numerical variables}
}
\value{
an object of \code{\link{pert_params-class}}
}
\description{
creates an \code{\link{pert_params-class}} object containing information
about perturbation parameters
}
\examples{
pTable <- ck_create_ptab(
  type = "destatis",
  D = 5,
  V = 3
)
params <- ck_create_pert_params(
  big_n = 17312941,
  small_n = 12,
  ptab = pTable,
  stab = ck_generate_stab(small_c = 12),
  mtab = c(0.6, 0.4, 0.2))
print(class(params))
}
\seealso{
\url{https://preview.tinyurl.com/yyleyg48}
}
