% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ck_create_ptab.R
\name{ck_create_ptab}
\alias{ck_create_ptab}
\title{Create a perturbation table}
\usage{
ck_create_ptab(D, V, type, js = 0, pstay = NULL, optim = 1,
  mono = TRUE, epsilon = 0.0000001, ptab_size = 70)
}
\arguments{
\item{D}{perturbation parameter for maximum perturbation (scalar or vector)}

\item{V}{perturbation parameter for variance (scalar)}

\item{type}{character specifying the format of the input table. Valid choices
are \code{"abs"} and \code{"destatis"}.}

\item{js}{treshold value for blocking of small frequencies (i.e. there
won't occur positive target frequencies below the treshold value)}

\item{pstay}{optional parameter to set}

\item{optim}{optimization parameter: \code{1} standard approach (default)}

\item{mono}{(logical) vector specifying optimization parameter for monotony
condition}

\item{epsilon}{(double)}

\item{ptab_size}{(number) defining the number of required columns in the
"abs"-approach}
}
\value{
an object of class \code{ptable} defined in the ptable-package.
}
\description{
This function allows to generate generates perturbation tables using
functionality from package "ptable".
}
\examples{
# "abs" style
ck_create_ptab(
  D = 5,
  V = 3,
  js = 2,
  ptab_size = 70,
  type = "abs"
)

# "destatis" style
ck_create_ptab(
  D = 5,
  V = 3,
  js = 2,
  pstay = 0.5,
  optim = 1,
  mono = TRUE,
  typ = "destatis"
)
}
\seealso{
This function uses functionality from package ptable
(https://github.com/sdcTools/ptable), expecially \code{pt_create_pParams}
and \code{pt_create_pTable}. More detailed information on the parameters
is available from the respective help-pages of these functions.
}
