% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ck_create_pTable.R
\name{ck_create_pTable}
\alias{ck_create_pTable}
\title{ck_create_pTable}
\usage{
ck_create_pTable(pTableSize = 75, ...)
}
\arguments{
\item{pTableSize}{(number) defining the number of required columns}

\item{...}{additional parameters (currently ignored)}
}
\value{
a \code{data.table} with 256 rows and the specified number of
columns given by parameter \code{pTableSize} holding perturbation values.
}
\description{
generates a simple perturbation table for developing/testing purposes.
Real-world lookup tables will be likely created in another package/software.
}
\examples{
ck_create_pTable(pTableSize=10)
}
