% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ck_create_testdata.R
\name{ck_create_testdata}
\alias{ck_create_testdata}
\title{ck_create_testdata}
\usage{
ck_create_testdata()
}
\value{
a \code{data.frame}
}
\description{
this function generates some test-data
}
\examples{
dat <- ck_create_testdata(); print(str(dat))
}
