% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ck_dat_hc92}
\alias{ck_dat_hc92}
\title{A real-world data set on persons}
\format{
ck_dat_hc92: a data frame with 820000 observations on the following 6 variables.
\itemize{
\item \code{id}: a numeric identifier
\item \code{geo_m}: a character vector defining regions
\item \verb{sex a}: character vector defining gender
\item \code{age_m}: a character vector containing age groups
\item \code{yae_h}: a character vector
\item \code{rkey}: a numeric vector holding record keys
}
}
\description{
820000 obervations in 5 Variables without sampling weights.
}
\examples{
data(ck_dat_hc92)
head(ck_dat_hc92)
}
\references{
https://ec.europa.eu/eurostat/cros/content/3-random-noise-cell-key-method_en
}
\keyword{datasets}
