% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ck_export_table.R
\name{ck_export_table}
\alias{ck_export_table}
\title{Method to extract information about modifications for count tables}
\usage{
ck_export_table(x, vname = NULL, type = "both")
}
\arguments{
\item{x}{input object}

\item{vname}{(character) variable name of a perturbed count variable or a continous variable for which the output
should be prepared}

\item{type}{(character) specify the type of output. Allowed values are:
\itemize{
\item \code{"both"}: unweighted and weighted original and perturbed counts/values are returned
\item \code{"unweighted"}: unweighted original and perturbed counts/values are returned
\item \code{"weighted"}: weighted and weighted original and perturbed counts/values are returned
}}
}
\value{
a \code{data.table} containing all dimensional variables and, depending
on the value of argument \code{type} and weather if \code{vname} is a count
or a continuously scaled variable, the following columns: For count-variables:
\itemize{
\item{UWC}: unweighted counts (argument \code{type} is "both" or "unweighted")
\item{WC}: perturbed unweighted counts (argument \code{type} is "both" or "unweighted")
\item{pUWC}: weighted counts (argument \code{type} is "both" or "weighted")
\item{pWC}: perturbed weighted counts (argument \code{type} is "both" or "weighted")
}
For continuously-scaled variables (magnitude tables):
\itemize{
\item{UW}: unweighted sum (argument \code{type} is "both" or "unweighted")
\item{WS}: perturbed unweighted sum (argument \code{type} is "both" or "unweighted")
\item{pUW}: weighted sum (argument \code{type} is "both" or "weighted")
\item{pWS}: perturbed weighted sum (argument \code{type} is "both" or "weighted")
}
}
\description{
Method to extract information about modifications for count tables
}
\examples{
## see example in ?perturbTable
}
