% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ck_freq_table.R
\name{ck_freq_table}
\alias{ck_freq_table}
\title{ck_freq_table}
\usage{
ck_freq_table(inp, vname = NULL)
}
\arguments{
\item{inp}{an object of class \code{\linkS4class{pert_table}} generated with \code{\link{perturbTable}}}

\item{vname}{either \code{NULL} or the name of a variable for which a perturbed
frequeny table has been generated. If \code{NULL}, the function prints the variables for
which a perturbation has been applied.}
}
\value{
a \code{data.table} containing all combinations of the dimensional variables in
the first n columns. Afterwards, the following columns are shown:
\itemize{
\item \code{UWC_{vname}}: unweighted counts for variable specified in \code{vname}
\item \code{WC_{vname}}: weighted counts for given variable
\item \code{pUWC_{vname}}: perturbed unweighted counts for given variable
\item \code{pWC_{vname}}: perturbed weighted counts for given variable
\item \code{WCavg_{vname}}: average cell weight for variable specified in \code{vname}
\item \code{row_indices}: row index in respect to the perturbation table used
\item \code{col_indices}: column index in respect to the perturbation table used. This
value is always \code{NA} in case "destatis"-type lookup tables have been used.
\item \code{pert}: actual amount of perturbation \code{vname}
\item \code{cellKey}: actual cellKey used in the lookup-procedure \code{vname}
}
}
\description{
this function allows to return perturbed count-tables from oututs of
\code{\linkS4class{pert_table}} objects typically generated
using \code{\link{perturbTable}}.
}
\examples{
## see examples in perturbTable()
}
\seealso{
ck_cont_table perturbTable
}
\author{
Bernhard Meindl
}
