% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ck_generate_rkeys.R
\name{ck_generate_rkeys}
\alias{ck_generate_rkeys}
\title{ck_generate_rkeys}
\usage{
ck_generate_rkeys(dat, max_val = (2^6) - 1)
}
\arguments{
\item{dat}{microdata used to generated hash for random seed}

\item{max_val}{maximum number for record-key}
}
\value{
a numeric vector with \code{N} record keys
}
\description{
creates random record keys by sampling between 1:max_val
}
\examples{
dat <- ck_create_testdata()
dat$rkeys <- ck_generate_rkeys(dat=dat, max_val=(2^6)-1)
}
