% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ck_generate_stab.R
\name{ck_generate_stab}
\alias{ck_generate_stab}
\title{Generate a perturbation table for numerical variables}
\usage{
ck_generate_stab(small_c = 12, ...)
}
\arguments{
\item{small_c}{number of columns for the perturbation table for large cells}

\item{...}{additional parameters, not yet used}
}
\value{
a \code{data.table} with 256 rows and 32 + \code{small_c} columns
}
\description{
This function allows to create a perturbation table (differently for for
large and small cell values) for testing purposes.
}
\examples{
ck_generate_stab(small_c = 10)
}
