% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ck_generate_sTable.R
\name{ck_generate_sTable}
\alias{ck_generate_sTable}
\title{ck_generate_sTable}
\usage{
ck_generate_sTable(smallC = 12, ...)
}
\arguments{
\item{smallC}{number of columns for the perturbation table for large cells}

\item{...}{additional parameters, not yet used}
}
\value{
a \code{data.table} with 256 rows and 32+\code{smallC} columns
}
\description{
a perturbation table (for large and small cells) for testing purposes
}
\examples{
ck_generate_sTable(smallC = 10)
}
