% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ck_params_nums.R
\name{ck_params_nums}
\alias{ck_params_nums}
\title{Create perturbation parameters for continuous variables}
\usage{
ck_params_nums(P = "top_contr", top_k = NULL, D, l, same_key = TRUE,
  use_zero_rkeys = FALSE)
}
\arguments{
\item{P}{a character value defining the way to identify the \code{magnifier},
e.g which contributions/values in a cell should be the used  in the
perturbation procedure. Possible choices are:
\itemize{
\item \code{top_contr}: the \code{k} largest contributions are used. In this case,
it is also required to specify argument \code{top_k}
\item \code{mean}: the (weighted) cellmean is used as starting point
\item \code{range}: the difference between largest and smallest contribution
is used.
\item \code{sum}: the (weighted) cellvalue itself is used as starting point
}}

\item{top_k}{it is ignored if \code{P} is different from \code{top_contr}. Otherwise,}

\item{D}{maximum perturbation value}

\item{l}{stepwidth parameter for computation of perturbation tables
a integerish number >= 1 specifying the number of top contributions whose
values should be perturbed.}

\item{same_key}{(logical) should original cell key (\code{TRUE}) used for
for finding perturbation values of the largest contributor to a
cell or should a perturbation of the cellkey itself (\code{FALSE}) take place.}

\item{use_zero_rkeys}{(logical) skalar defining if record keys of
units not contributing to a specific numeric variables should be
used (\code{TRUE}) or ignored (\code{FALSE}) when computing cell keys.}
}
\value{
an object suitable as input to \code{\link[=ck_setup]{ck_setup()}} for the perturbation
of magnitude tables
}
\description{
This function allows to define perturbation parameters used to
perturb cells in magnitude tables.
}
\examples{
ck_params_nums(D = 3, l = 0.5, P = "top_contr", top_k = 3)
ck_params_nums(D = 10, l = .5, P = "mean")
}
