% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ck_create_custom_pTable.R
\name{ck_update_custom_pTable}
\alias{ck_update_custom_pTable}
\title{Modify a custom perturbation table}
\usage{
ck_update_custom_pTable(pTable, fun, cols = NULL, rows = NULL)
}
\arguments{
\item{pTable}{an object created using \code{\link{ck_create_custom_pTable}}}

\item{fun}{a function that will be used for given cells to derive the
perturbation values. The function must return exactly one number when called
without parameters.}

\item{cols}{numeric index specifying columns for which the given perturbation
function should be set. If \code{NULL}, the function will be set on all columns
of the perturbation table.}

\item{rows}{numeric index specifying rows for which the given perturbation
function should be set. If \code{NULL}, the function will be set on all
rows of the perturbation table.}
}
\value{
an object of class \code{ptable} defined in the ptable-package.
}
\description{
Modify a custom perturbation table
}
\examples{
## see examples in ?ck_create_custom_pTable and ?perturbTable
}
\author{
Bernhard Meindl
}
