% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ck_create_free_ptable.R
\name{ck_update_free_ptable}
\alias{ck_update_free_ptable}
\title{Modify a perturbation table in "free" format}
\usage{
ck_update_free_ptable(ptab, fun, cols = NULL, rows = NULL)
}
\arguments{
\item{ptab}{an object created using \code{\link{ck_create_free_ptable}}}

\item{fun}{a function that will be used for given cells to derive the
perturbation values. The function must return exactly one number when called
without parameters.}

\item{cols}{numeric index specifying columns for which the given perturbation
function should be set. If \code{NULL}, the function will be set on all
columns of the perturbation table.}

\item{rows}{numeric index specifying rows for which the given perturbation
function should be set. If \code{NULL}, the function will be set on all
rows of the perturbation table.}
}
\value{
an object of class \code{ptable} defined in the ptable-package.
}
\description{
This function allows to change and update a perturbation table that
is in "free" format.
}
\examples{
## see examples in ?ck_create_free_ptable and ?perturb_table
}
\author{
Bernhard Meindl
}
