% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ck_vignette.R
\name{ck_vignette}
\alias{ck_vignette}
\title{ck_vignette}
\usage{
ck_vignette()
}
\value{
a browser windows/tab with showing the vignette
}
\description{
starts the package vignette that gets you started with the package
}
\examples{
\dontrun{
ck_vignette()
}
}
