% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{mdata}
\alias{mdata}
\title{A real-world data set on persons}
\format{mdata: a data frame with 820000 observations on the following 5 variables.
\describe{
\item{geo_m}{a character vector defining regions}
\item{sex}{a character vector defining gender}
\item{age_m}{a character vector containing age groups}
\item{yae_h}{a character vector}
\item{rkey}{a numeric vector holding record keys}}}
\description{
820000 obervations in 5 Variables without sampling weights.
}
\examples{
data(mdata)
head(mdata)
}
\references{
https://ec.europa.eu/eurostat/cros/content/3-random-noise-cell-key-method_en
}
\keyword{datasets}
