% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ck_methods.R
\docType{methods}
\name{mod_counts}
\alias{mod_counts}
\alias{mod_counts,pert_table-method}
\title{Method to extract information about modifications for count tables}
\usage{
mod_counts(x, ...)

\S4method{mod_counts}{pert_table}(x, verbose = FALSE)
}
\arguments{
\item{x}{input object}

\item{...}{additional parameters passed to specific methods}

\item{verbose}{(logical) if \code{TRUE}, additional information is printed}
}
\description{
Method to extract information about modifications for count tables
}
\examples{
## see example in ?perturbTable
}
