% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ck_methods.R
\docType{methods}
\name{mod_numvars}
\alias{mod_numvars}
\alias{mod_numvars,pert_table-method}
\title{Method to extract information about modifications for counts}
\usage{
mod_numvars(x, ...)

\S4method{mod_numvars}{pert_table}(x)
}
\arguments{
\item{x}{input object}

\item{...}{additional parameters passed to methods}
}
\description{
Method to extract information about modifications for counts
}
