% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{params_cnts}
\alias{params_cnts}
\title{Perturbation parameters for count variables}
\format{params_cnts: a \code{ck_params} object that can be used as input in \code{\link[=cellkey_pkg]{cellkey_pkg()}} and
contains parameters to perturb count variables}
\description{
This object is a result from [ck_params_cnts())] and is used in unit-tests.

[ck_params_cnts())]: R:ck_params_cnts())
}
\examples{
str(params_cnts)
}
\references{
https://ec.europa.eu/eurostat/cros/content/3-random-noise-cell-key-method_en
}
\keyword{internal}
