% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ck_classes.R
\docType{class}
\name{pert_inputdat-class}
\alias{pert_inputdat-class}
\title{An S4-class to represent input data for applying the cell-key method}
\description{
An S4-class to represent input data for applying the cell-key method
}
\section{Slots}{

\describe{
\item{\code{microdat}}{data.table containing microdata}

\item{\code{rkeys}}{(integer) vector specifying record keys}

\item{\code{pert_params}}{pert_params information about perturbation parameters}
}}

