% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ck_classes.R, R/ck_methods.R
\docType{class}
\name{pert_table-class}
\alias{pert_table-class}
\alias{results,pert_table-method}
\alias{print,pert_table-method}
\alias{mod_counts,pert_table-method}
\alias{mod_numvars,pert_table-method}
\title{An S4 class to represent a perturbed table}
\usage{
\S4method{results}{pert_table}(x, meanBeforeSum = TRUE)

\S4method{print}{pert_table}(x)

\S4method{mod_counts}{pert_table}(x)

\S4method{mod_numvars}{pert_table}(x)
}
\arguments{
\item{x}{an object of class \code{\link{pert_table-class}}}

\item{meanBeforeSum}{(logical) defines if for (continuous) variables the weighted means should be computed
before weighted sums or the other way round.}
}
\description{
An S4 class to represent a perturbed table
}
\section{Slots}{

\describe{
\item{\code{tab}}{a \code{data.table} containing original and pertubed values. The following variables are always present:
\itemize{
\item \strong{UWC: } unweighted counts
\item \strong{pUWC: } perturbed unweighted counts
\item \strong{WC: } weighted counts
\item \strong{pWC: } perturbed weighted counts
\item \strong{WCavg: } average weight for each cell
}
Additionally, for each numerical variable the perturbed variable named {vName.pert} is included in this table.}

\item{\code{count_modifications}}{a \code{data.table} with 3 columns (\code{row_indices}, \code{col_indices}
and \code{applied_perturbation}) that contains information for each cell, where the applied perturbation
was extracted from the perturbation table (useful for debugging)}

\item{\code{numvars_modifications}}{(list) containing for each numerical variable
that was tabulated a list element with a \code{data.table} showing which
values have been modified prior to tabulation.}

\item{\code{numVars}}{(character) variable names of numeric variables that have been tabulated}

\item{\code{cellKeys}}{(integer) vector containing the cell keys}

\item{\code{is_weighted}}{(logical) TRUE if sampling weights have been used}
}}

\seealso{
\url{https://www.unece.org/fileadmin/DAM/stats/documents/ece/ces/ge.46/2013/Topic_1_ABS.pdf}
}
