% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perturb_table.R
\name{perturb_table}
\alias{perturb_table}
\title{Perturb a multi-dimensional table using the cell-key method}
\usage{
perturb_table(inp, dim_list, countVars = NULL, numVars = NULL,
  by = NULL, weightVar = NULL)
}
\arguments{
\item{inp}{an object of class \code{\link{pert_inputdat-class}}
generated with \code{\link{ck_create_input}}}

\item{dim_list}{a list containing slots for each variable that should be
tabulated. Each slot consists of a \code{data.frame} or \code{data.table}
with columns \code{levels} and \code{codes} specifiying the
hierarchies of the variables.}

\item{countVars}{(character) vector of numerical variables that should be
counted and perturbed; must be 0/1 coded.}

\item{numVars}{(character) vector of numerical variables that
should be tabulated or \code{NULL}}

\item{by}{\code{NULL} or a scalar character. If specified, it is possible
to use a valid variable name being 0/1 coded that will be used to create the
perturbation of all variables specified in \code{numVars} only for subgroups
defined by the specified count variable. If the variable is not listed in
argument \code{countVars} it will be automatically added. If not specified or
\code{NULL}, the table will be based on all units in the microdata.}

\item{weightVar}{(character) vector of variable holding sampling
weights or \code{NULL}}
}
\value{
an object of class \code{\link{pert_table-class}}.
}
\description{
This function allows to compute perturbed tables containing counts as
well as continous variables.
}
\examples{
# loading testdata and adding record keys
dat <- ck_create_testdata()

# create some binary (0,1) variables that should be perturbed later
dat[, cnt_females := ifelse(sex == "male", 0, 1)]
dat[, cnt_males := ifelse(sex == "male", 1, 0)]
dat[, cnt_highincome := ifelse(income >= 9000, 1, 0)]

## create record keys
dat$rkeys <- ck_generate_rkeys(
  dat = dat,
  max_val = 2*nrow(dat),
  type = "abs",
  verbose = TRUE
)

# creating test perturbation in "destatis", "abs" and "free" format
ptab_abs <- ck_create_ptab(
  D = 5,
  V = 3,
  pTableSize = 70,
  type = "abs"
)
ptab_destatis <- ck_create_ptab(
  D = 5,
  V = 3,
  type = "destatis"
)

# "free" format is a bit more complicated to set up
ptab_free <- ck_create_free_ptable(
  nrcol = 70
)

# create some functions first that return a single number
fn1 <- function() round(rnorm(1, mean = 5, sd = 10))
fn2 <- function() rpois(1, lambda = 2)
fn3 <- function() rpois(1, lambda = 5)
fn4 <- function() return(1)
fn5 <- function() return(-1)

# fn1 provides perturbation values from a normal distribution
# with mean = 5 and sd = 10; we use this for all cells
ptab_free <- ck_update_free_ptable(
  ptab = ptab_free,
  fun = fn1
)

# perturbation values from poisson distributions for some cells
ptab_free <- ck_update_free_ptable(
  ptab = ptab_free,
  fun = fn2,
  cols = 1:20,
  rows = 1:100
)
ptab_free <- ck_update_free_ptable(
  ptab = ptab_free,
  fun = fn3,
  cols = 1:20,
  rows = 101:256
)

# we can of course write functions, that return scalars, such
# as fn4() and fn5() always return 1 or -1
ptab_free <- ck_update_free_ptable(
  ptab = ptab_free,
  fun = fn4,
  cols = 10:20
)
ptab_free <- ck_update_free_ptable(
  ptab = ptab_free,
  fun = fn5,
  cols = 21:30
)

# set up perturbation tables and parameters for numerical variables
stab <- ck_generate_stab(small_c = 12)
mtab <- c(0.6, 0.4, 0.2)

# creating perturbation objects that are later used
# as inputs in perturb_table()
pert_params_abs <- ck_create_pert_params(
  big_n = 17312941,
  small_n = 12,
  ptab = ptab_abs,
  stab = stab,
  mtab = mtab
)

pert_params_destatis <- ck_create_pert_params(
  small_n = 12,
  ptab = ptab_destatis,
  stab = stab,
  mtab = mtab
)

pert_params_free <- ck_create_pert_params(
  big_n = 17312941,
  small_n = 12,
  ptab = ptab_free,
  stab = stab,
  mtab = mtab
)

# create suitable input data using existing record keys
# and a perturbation table in "abs" format"
inp_abs <- ck_create_input(
  dat = dat,
  def_rkey = "rkeys",
  pert_params = pert_params_abs
)

# create input using a perturbation table in "destatis" format"
# and no existing record keys. In this case, def_rkey = 9 means
# that the generated record keys are rounded to 9 digits
inp_destatis <- ck_create_input(
  dat = dat,
  def_rkey = 9,
  pert_params = pert_params_destatis
)

# input using a perturbation table in "free" format
# consisting making use of custom functions
inp_free <- ck_create_input(
  dat = dat,
  def_rkey = "rkeys",
  pert_params = pert_params_free
)

# define table hierarchies
# variable "sex"
dim_sex <- hier_create(root = "Total", nodes = c("male", "female"))
hier_display(dim_sex)

# variable "age"
dim_age <- hier_create(root = "Total", nodes = paste0("age_group", 1:6))
hier_display(dim_age)

# define required inputs
dim_list <- list(sex = dim_sex, age = dim_age)
weightVar <- "sampling_weight"
numVars <- c("savings", "income")

# finally perturb the table
res <- perturb_table(
  inp = inp_abs,
  dim_list = dim_list,
  weightVar = weightVar,
  countVars = NULL,
  numVars = numVars
)

# access the results
ck_freq_table(res, vname = "Total")
ck_cont_table(res, vname = "income", mean_before_sum = TRUE)
ck_cont_table(res, vname = "savings", mean_before_sum = TRUE)

res <- perturb_table(
  inp = inp_destatis,
  dim_list = dim_list,
  weightVar = weightVar,
  countVars = NULL,
  numVars = numVars
)
ck_freq_table(res, vname = "Total")
ck_cont_table(res, vname = "income", mean_before_sum = TRUE)
ck_cont_table(res, vname = "savings", mean_before_sum = TRUE)

res <- perturb_table(
  inp = inp_free,
  dim_list = dim_list,
  weightVar = weightVar,
  countVars = NULL,
  numVars = numVars
)
ck_freq_table(res, vname = "Total")
ck_cont_table(res, vname = "income", mean_before_sum = TRUE)
ck_cont_table(res, vname = "savings", mean_before_sum = TRUE)

# information on modifications for count variables
mod_counts(res, verbose = TRUE)

# information on modifications for numerical variables
mod_numvars(res)

# measures of information loss and utility statistics
# on tabulated count variables
ck_cnt_measures(res, vname = "Total")

# an example using additional countVars
res <- perturb_table(
  inp = inp_destatis,
  dim_list = dim_list,
  weightVar = weightVar,
  countVars = c("cnt_females", "cnt_males", "cnt_highincome"),
  numVars = numVars)
print(res) # custom print method
summary(res) # custom summary method

# show count tables
ck_freq_table(res, vname = NULL)
ck_freq_table(res, vname = "Total")
ck_freq_table(res, vname = "cnt_females")
ck_freq_table(res, vname = "cnt_males")
ck_freq_table(res, vname = "cnt_highincome")

# show magnitude tables along with its modifications
p_inc <- ck_cont_table(res, vname = "income"); p_inc
attr(p_inc, "modifications")

# create perturbed magnitude table for subgroups of data
res <- perturb_table(
  inp = inp_destatis,
  dim_list = dim_list,
  weightVar = weightVar,
  countVars = c("cnt_females", "cnt_males", "cnt_highincome"),
  numVars = numVars,
  by = "cnt_males"
)

# the perturbed tables of savings and income are computed only where
# variable cnt_males is 1
p_sav <- ck_cont_table(res, vname = "savings"); p_sav
attr(p_sav, "modifications") # no modifications in cells containing females!

# export table to a simple data.table
df_tot <- ck_export_table(res, vname = "Total", type = "both")
head(df_tot)
df_inc <- ck_export_table(res, vname = "income", type = "weighted")
head(df_inc)
}
\seealso{
\url{https://preview.tinyurl.com/yyleyg48}
}
