% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perturbTable.R
\name{perturbTable}
\alias{perturbTable}
\title{perturbTable}
\usage{
perturbTable(inp, dimList, numVars = NULL, weightVar = NULL)
}
\arguments{
\item{inp}{an object of class \code{\link{pert_inputdat-class}} generated with \code{\link{ck_create_input}}}

\item{dimList}{a list containing slots for each variable that should be tabulated. Each slot consists of a \code{data.frame} or \code{data.table}
with columns \code{levels} and \code{codes} specifiying the hierarchies of the variables.}

\item{numVars}{(character) vector of numerical variables that should be tabulated or \code{NULL}}

\item{weightVar}{(character) vector of variable holding sampling weights or \code{NULL}}
}
\value{
an object of class \code{\link{pert_table-class}}.
}
\description{
compute perturbed tables containing counts and continous variables.
}
\examples{
dat <- ck_create_testdata()
dat$age <- as.integer(cut(dat$age, 6))
maxV <- 15*nrow(dat)
dat$rkeys <- ck_generate_rkeys(dat=dat, max_val=maxV)
pert_params <- ck_create_pert_params(
  bigN=17312941, smallN=12,
  pTable=ck_create_pTable(pTableSize=70),
  sTable=ck_generate_sTable(smallC=12),
  mTable=c(0.6,0.4,0.2))

inp <- ck_create_input(dat=dat, def_rkey="rkeys", pert_params=pert_params)

dim.sex <- data.table(levels=c("@","@@","@@"), codes=c("Total", 1, 2))
dim.age <- data.table(levels=c("@",rep("@@", 6)), codes=c("Total", 1:6))
dimList <- list(sex=dim.sex, age=dim.age)
res <- perturbTable(inp=inp, dimList=dimList,
  weightVar="sampling_weight",
  numVars=c("savings", "income"))
results(res, meanBeforeSum=TRUE)

## information on modifications for count variables
mod_counts(res)

## information on modifications for numerical variables
mod_numvars(res)
}
\seealso{
\url{https://www.unece.org/fileadmin/DAM/stats/documents/ece/ces/ge.46/2013/Topic_1_ABS.pdf}
}
