% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ck_methods.R
\docType{methods}
\name{print,pert_table-method}
\alias{print,pert_table-method}
\title{Method to extract results}
\usage{
\S4method{print}{pert_table}(x)
}
\arguments{
\item{x}{input object}
}
\description{
Method to extract results
}
