% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ptable_destatis}
\alias{ptable_destatis}
\title{A perturbation table based on DESTATIS format}
\format{ptable_destatis: a data frame with 88 rows on the following
variables.
\describe{
\item{i}{a numeric vector}
\item{j}{a numeric vector}
\item{p}{a numeric vector}
\item{kum_p_u}{a numeric vector}
\item{kum_p_o}{a numeric vector}
\item{diff}{a numeric vector}}}
\description{
66 rows in 6 Variables
}
\examples{
data(ptable_destatis)
head(ptable_destatis)
}
\keyword{datasets}
