% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ck_methods.R
\name{results}
\alias{results}
\title{Method to extract results}
\usage{
results(x, ...)
}
\arguments{
\item{x}{input object}

\item{...}{additional parameters passed to methods}
}
\description{
Method to extract results
}
