% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{testdata}
\alias{testdata}
\title{A real-world data set on household income and expenditures}
\format{testdata: a data frame with 4580 observations on the following
15 variables.
\describe{
\item{urbrur}{a numeric vector}
\item{roof}{a numeric vector}
\item{walls}{a numeric vector}
\item{water}{a numeric vector}
\item{electcon}{a numeric vector}
\item{relat}{a numeric vector}
\item{sex}{a numeric vector}
\item{age}{a numeric vector}
\item{hhcivil}{a numeric vector}
\item{expend}{a numeric vector}
\item{income}{a numeric vector}
\item{savings}{a numeric vector}
\item{ori_hid}{a numeric vector}
\item{sampling_weight}{a numeric vector}
\item{household_weights}{a numeric vector}}}
\description{
4580 Obervations in 15 Variables; This dataset also contains
sampling weights.
}
\examples{
data(testdata)
head(testdata)
}
\references{
The International Household Survey Network, www.ihsn.org
}
\keyword{datasets}
