% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ck_create_input.R
\name{ck_create_input}
\alias{ck_create_input}
\title{Create input for \code{\link{perturb_table}}}
\usage{
ck_create_input(dat, def_rkey, pert_params)
}
\arguments{
\item{dat}{a \code{data.table} containing micro data}

\item{def_rkey}{either a column name within \code{dat} specifying a variable
containing record keys or a single (integer) number.

If it is a number, we have to distinguish two cases:

\itemize{
\item The perturbation table is provided in "abs"-format:

\code{def_rkey} is used as upper bound to compute record keys between
1 and \code{def_rkey}
\item The perturbation table is provided in "destatis"-format:

\code{def_rkey} is used to specify the maximum number of digits for
record keys that are generated from a uniform distribution between 0 and 1.
}}

\item{pert_params}{information on perturbation parameters created
using \code{\link{ck_create_pert_params}}}
}
\value{
an object of class \code{\link{pert_inputdat-class}}
}
\description{
This function creates the required input for \code{\link{perturb_table}}.
}
\examples{
## define parameters
big_n <- 20011
small_n <- 12
maxV <- 214748365 # a large prime number
dat <- ck_create_testdata()
dat$rkeys <- ck_generate_rkeys(dat = dat, max_val = maxV)
mtab <- c(0.6,0.4,0.2)
stab <- ck_generate_stab(small_c = 12)
ptab <- ck_create_ptab(
  D = 5,
  V = 3,
  pTableSize = 70,
  type = "abs"
)

pert_params <- ck_create_pert_params(
  big_n = big_n,
  small_n = small_n,
  ptab = ptab,
  stab = stab,
  mtab = mtab
)

## create suitable input data using exising record keys
inp <- ck_create_input(
  dat = dat,
  def_rkey = "rkeys",
  pert_params = pert_params
)
print(class(inp))
}
