% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ck_create_pert_params.R
\name{ck_create_pert_params}
\alias{ck_create_pert_params}
\title{ck_create_pert_params}
\usage{
ck_create_pert_params(bigN, smallN, pTable, sTable, mTable)
}
\arguments{
\item{bigN}{(integer) large prime number used to derive cell keys from record keys}

\item{smallN}{(integer) parameter that controlls how the columns of \code{pTable} are scrolled}

\item{pTable}{a perturbation table generated with \code{\link{ck_create_pTable}}}

\item{sTable}{table containing the noise component for numerical variables generated by 
\code{\link{ck_generate_sTable}}}

\item{mTable}{(numeric) vector containing parameters for the magnitude of perturation for numerical variables}
}
\value{
an object of \code{\link{pert_params-class}}
}
\description{
creates an \code{\link{pert_params-class}} object containing information
about perturbation parameters
}
\examples{
params <- ck_create_pert_params(
  bigN=17312941, 
  smallN=12, 
  pTable=ck_create_pTable(pTableSize=70),
  sTable=ck_generate_sTable(smallC=12),
  mTable=c(0.6,0.4,0.2))
print(class(params))
}
\seealso{
\url{https://www.unece.org/fileadmin/DAM/stats/documents/ece/ces/ge.46/2013/Topic_1_ABS.pdf}
}
