% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ck_params_nums.R
\name{ck_flexparams}
\alias{ck_flexparams}
\title{Define parameters for numeric perturbation magnitudes using a flex function}
\usage{
ck_flexparams(flexpoint, m_small = 0.25, m_large = 0.05, epsilon = 1,
  scaling = TRUE, q = 3)
}
\arguments{
\item{flexpoint}{(numeric scalar); at which point should the noise coefficient
function reaches its desired maximum (defined by \code{m_small})}

\item{m_small}{(numeric scalar); the desired maximum percentage for the function (for small values)}

\item{m_large}{(numeric scalar); the desired noise percentage for larger values}

\item{epsilon}{a numeric vector in descending order with all values \code{>= 0} and \code{<= 1} with the first
element forced to equal 1. The length of this vector must correspond with the number \code{topK}
specified in \code{\link[=ck_params_nums]{ck_params_nums()}} when creating parameters for \code{type == "top_contr"} which is
checked at runtime. This setting allows to use different flex-functions for the largest \code{top_k} contributors.}

\item{scaling}{(logical scalar); if \code{TRUE}, the magnitude parameter is computed as described in
deliverable 4.2 (section \code{2.3.1}); otherwise a simpler approach is taken where the perturbation magnitudes
are proportional to \code{epsilon} times \code{m_large} for sufficient large values.}

\item{q}{(numeric scalar); Parameter of the function; \code{q} needs to be \code{>= 1}}
}
\value{
an object suitable as input for \code{\link[=ck_params_nums]{ck_params_nums()}}.
}
\description{
\code{\link[=ck_flexparams]{ck_flexparams()}} allows to define a flex function that is used to lookup perturbation
magnitudes (percentages) used when perturbing continuous variables.
}
\details{
details about the grid function can be found in Deliverable D4.2, Part I in
SGA \emph{"Open Source tools for perturbative confidentiality methods"}
}
\examples{
x <- ck_create_testdata()

# create some 0/1 variables that should be perturbed later
x[, cnt_females := ifelse(sex == "male", 0, 1)]
x[, cnt_males := ifelse(sex == "male", 1, 0)]
x[, cnt_highincome := ifelse(income >= 9000, 1, 0)]
# a variable with positive and negative contributions
x[, mixed := sample(-10:10, nrow(x), replace = TRUE)]

# create record keys
x$rkey <- ck_generate_rkeys(dat = x)

# define required inputs

# hierarchy with some bogus codes
d_sex <- hier_create(root = "Total", nodes = c("male", "female"))
d_sex <- hier_add(d_sex, root = "female", "f")
d_sex <- hier_add(d_sex, root = "male", "m")

d_age <- hier_create(root = "Total", nodes = paste0("age_group", 1:6))
d_age <- hier_add(d_age, root = "age_group1", "ag1a")
d_age <- hier_add(d_age, root = "age_group2", "ag2a")

# define the cell key object
countvars <- c("cnt_females", "cnt_males", "cnt_highincome")
numvars <- c("expend", "income", "savings", "mixed")
tab <- ck_setup(
  x = x,
  rkey = "rkey",
  dims = list(sex = d_sex, age = d_age),
  w = "sampling_weight",
  countvars = countvars,
  numvars = numvars)

# show some information about this table instance
tab$print() # identical with print(tab)

# which variables have been defined?
tab$allvars()

# count-variables
tab$cntvars()

# continuous variables
tab$numvars()

# add perturbation parameters for "total" variable
p_cnts1 <- ck_params_cnts(
  D = 5,
  V = 3,
  js = 2,
  pstay = 0.5,
  optim = 1,
  mono = TRUE)
tab$params_cnts_set(val = p_cnts1, v = "total")

# create alternative perturbation parameters
p_cnts2 <- ck_params_cnts(
  D = 8,
  V = 3,
  js = 2,
  pstay = 0.5,
  optim = 1,
  mono = TRUE)

# use it for the remaining variables
tab$params_cnts_set(val = p_cnts2, v = countvars)

# perturb a variable
tab$perturb(v = "total")

# multiple variables can be perturbed as well
tab$perturb(v = c("cnt_males", "cnt_highincome"))

# return results (weighted and unweighted)
tab$freqtab(v = c("total", "cnt_males"))

# only unweighted and perturbed unweighted counts
tab$freqtab(v = c("total", "cnt_males"), type = "unweighted")

# numerical variables (positive variables using flex-function)
p_nums1 <- ck_params_nums(
  D = 10,
  l = 0.5,
  type = "top_contr",
  top_k = 3,
  mult_params = ck_flexparams(
    flexpoint = 1000,
    m_small = 0.30,
    m_large = 0.03,
    epsilon = c(1, 0.5, 0.2),
    q = 3
  ),
  mu_c = 2,
  m_fixed_sq = 4,
  same_key = FALSE,
  pos_neg_var = 0
)

# another set of parameters
# for variables with positive and negative values
p_nums2 <- ck_params_nums(
  D = 10,
  l = 0.5,
  type = "top_contr",
  top_k = 3,
  mult_params = ck_flexparams(
    flexpoint = 1000,
    m_small = 0.15,
    m_large = 0.02,
    epsilon = c(1, 0.4, 0.15),
    q = 3,
    scaling = FALSE
  ),
  mu_c = 2,
  m_fixed_sq = 4,
  same_key = FALSE,
  pos_neg_var = 1
)

# use `p_nums1` for all variables
tab$params_nums_set(p_nums1, c("savings", "income", "expend"))

# use different parameters for variable `mixed`
tab$params_nums_set(p_nums2, v = "mixed")

# perturb variables
tab$perturb(v = c("income", "savings"))

# extract results
tab$numtab("income", mean_before_sum = TRUE)
tab$numtab("income", mean_before_sum = FALSE)
tab$numtab("savings")

# results can be resetted, too
tab$reset_cntvars(v = "cnt_males")

# we can then set other parameters and perturb again
tab$params_cnts_set(val = p_cnts1, v = "cnt_males")
tab$perturb(v = "cnt_males")

# write to a file "outtab.csv" (.csv is automatically added to the path)
\dontrun{
tab$freqtab(v = c("total", "cnt_males"), path = "outtab")
}

# only weighted and perturbed weighted counts
tab$freqtab(v = c("total", "cnt_males"), type = "weighted")

# utility measures for a count variable
tab$measures_cnts(v = "total", exclude_zeros = TRUE)

# modifications for perturbed count variables
tab$mod_cnts()

# display a summary about utility measures
tab$summary()
}
\seealso{
\code{\link[=ck_params_nums]{ck_params_nums()}}
}
