% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ck_params_nums.R
\name{ck_params_nums}
\alias{ck_params_nums}
\title{Set perturbation parameters for continuous variables}
\usage{
ck_params_nums(type = "top_contr", top_k = NULL, D, l, mult_params,
  mu_c = 0, same_key = TRUE, use_zero_rkeys = FALSE,
  pos_neg_var = 1)
}
\arguments{
\item{type}{a character value defining the way to identify the \code{magnifier},
e.g which contributions/values in a cell should be the used in the
perturbation procedure. Possible choices are:
\itemize{
\item \code{top_contr}: the \code{k} largest contributions are used. In this case,
it is also required to specify argument \code{top_k}
\item \code{mean}: the (weighted) cellmean is used as starting point
\item \code{range}: the difference between largest and smallest contribution
is used.
\item \code{sum}: the (weighted) cellvalue itself is used as starting point
}}

\item{top_k}{it is ignored if \code{variant} is different from \code{top_contr}. Otherwise,}

\item{D}{maximum perturbation value}

\item{l}{stepwidth parameter for computation of perturbation tables
a integerish number >= 1 specifying the number of top contributions whose
values should be perturbed.}

\item{mult_params}{an object derived with \code{\link[=ck_flexparams]{ck_flexparams()}} or \code{\link[=ck_simpleparams]{ck_simpleparams()}}
that contain required parameters for the computation of the perturbation multiplier}

\item{mu_c}{fixed extra protection amount (\code{>= 0)} applied to the absolute of the
perturbation value of the first (largest) noise component; defaults to
\code{0} (no additional protection)}

\item{same_key}{(logical) should original cell key (\code{TRUE}) used for
for finding perturbation values of the largest contributor to a
cell or should a perturbation of the cellkey itself (\code{FALSE}) take place.}

\item{use_zero_rkeys}{(logical) skalar defining if record keys of
units not contributing to a specific numeric variables should be
used (\code{TRUE}) or ignored (\code{FALSE}) when computing cell keys.}

\item{pos_neg_var}{a number defining the strategy to look up perturbation values in case
the observations can take positive and negative values. This setting is ignored if the variable
has no negative values. The possible settings for parameter \code{pos_neg_var} are:
\itemize{
\item \code{0}: the variable must be strictly positive
\item \code{1}: the perturbation value is always selected from the block of perturbation
values referring to the symmetric case (\code{i} equals \code{D} in the perturbation table) independent
of the actual value of the observation
\item \code{2}: the lookup of a perturbation value \code{v} for a value \code{x} is done like in the case
where variables only take positive values using \code{abs(x)} to find the relevant block
in the perturbation table. The perturbed value is then computed as \code{sign(x) * (abs(x) + v)}.
\item \code{3}:    use variant \code{2} for all \code{x != 0} and apply \code{1} for \code{x == 0}
}}
}
\value{
an object suitable as input to method \code{$params_nums_set()} for the perturbation
of continous variables.
}
\description{
This function allows to define perturbation parameters used to
perturb cells in magnitude tables.
}
\examples{
ck_params_nums(
  D = 3,
  l = 0.5,
  type = "top_contr",
  top_k = 3,
  mult_params = ck_flexparams(
    fp = 1000,
    p = c(0.20, 0.03),
    epsilon = c(1, 0.5, 0.2),
    q = 2
  ),
  use_zero_rkeys = TRUE,
  mu_c = 3,
  pos_neg_var = 2
)
}
\seealso{
\code{\link[=ck_flexparams]{ck_flexparams()}}
}
