% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{perturbTable}
\alias{perturbTable}
\alias{ck_create_node}
\alias{ck_add_nodes}
\alias{ck_delete_nodes}
\alias{rename_node}
\title{Perturb a table}
\usage{
perturbTable(...)

ck_create_node(...)

ck_add_nodes(...)

ck_delete_nodes(...)

rename_node(...)
}
\description{
perturbTable() is defunct, please use
\code{\ļink{perturb_table}} instead

ck_create_node() is defunct, please use
sdcHierarchies::hier_create() instead

ck_add_nodes() is defunct, please use
sdcHierarchies::hier_add() instead

ck_delete_nodes() is defunct, please use
sdcHierarchies::hier_delete() instead

`rename_node()` is defunct, please use sdcHierarchies::hier_rename() instead
}
\keyword{internal}
