% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ck_methods.R
\name{mod_numvars}
\alias{mod_numvars}
\title{Method to extract information about modifications for counts}
\usage{
mod_numvars(x, ...)
}
\arguments{
\item{x}{input object}

\item{...}{additional parameters passed to methods}
}
\description{
Method to extract information about modifications for counts
}
