% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perturbTable.R
\name{perturbTable}
\alias{perturbTable}
\title{perturbTable}
\usage{
perturbTable(inp, dimList, numVars = NULL, weightVar = NULL)
}
\arguments{
\item{inp}{an object of class \code{\link{pert_inputdat-class}}
generated with \code{\link{ck_create_input}}}

\item{dimList}{a list containing slots for each variable that should be
tabulated. Each slot consists of a \code{data.frame} or \code{data.table}
with columns \code{levels} and \code{codes} specifiying the
hierarchies of the variables.}

\item{numVars}{(character) vector of numerical variables that
should be tabulated or \code{NULL}}

\item{weightVar}{(character) vector of variable holding sampling
weights or \code{NULL}}
}
\value{
an object of class \code{\link{pert_table-class}}.
}
\description{
compute perturbed tables containing counts and continous variables.
}
\examples{
## loading testdata and adding record keys (for abs-algorithm)
dat <- ck_create_testdata()
dat$rkeys <- ck_generate_rkeys(dat=dat, max_val=2*nrow(dat), type="abs", verbose=TRUE)

## loading test perturbation in abs- and destatis format
pTable_abs <- ck_create_pTable(pTableSize=70, type="abs")
pTable_destatis <- ck_create_pTable(type="destatis")

## perturbation tables and parameters for numerical variables
sTable <- ck_generate_sTable(smallC=12)
mTable <- c(0.6,0.4,0.2)

## creating perturbation objects that are later used
## as inputs in perturbTable()
pert_params_abs <- ck_create_pert_params(
  bigN=17312941, smallN=12,
  pTable=pTable_abs,
  sTable=sTable,
  mTable=mTable)

pert_params_destatis <- ck_create_pert_params(
  smallN=12,
  pTable=pTable_destatis,
  sTable=sTable,
  mTable=mTable)

## create suitable input data using existing record keys
## and a perturbation table in "abs-format"
inp_abs <- ck_create_input(
  dat=dat,
  def_rkey="rkeys",
  pert_params=pert_params_abs)

## create input using a perturbation table in "destatis-format"
## and no existing record keys. In this case, def_rkey=9 means
## that the generated record keys are rounded to 9 digits
inp_destatis <- ck_create_input(
  dat=dat,
  def_rkey=9,
  pert_params=pert_params_destatis)

## definining table hierarchies
## variable 'sex'
dim.sex <- ck_create_node(total_lab="Total")
dim.sex <- ck_add_nodes(dim.sex, reference_node="Total", node_labs=c("male","female"))
print(dim.sex)

## variable 'sex'
dim.age <- ck_create_node(total_lab="Total")
dim.age <- ck_add_nodes(dim.age,
  reference_node="Total",
  node_labs=paste0("age_group",1:6))
print(dim.age)

## define required inputs
dimList <- list(sex=dim.sex, age=dim.age)
weightVar <- "sampling_weight"
numVars <- c("savings", "income")

## perturbing the table
res_abs <- perturbTable(
  inp=inp_abs,
  dimList=dimList,
  weightVar=weightVar, numVars=numVars)
results(res_abs, meanBeforeSum=TRUE)

res_destatis <- perturbTable(
  inp=inp_destatis,
  dimList=dimList,
  weightVar=weightVar, numVars=numVars)
results(res_destatis, meanBeforeSum=TRUE)

## information on modifications for count variables
mod_counts(res_abs, verbose=TRUE)

## information on modifications for numerical variables
mod_numvars(res_abs)
}
\seealso{
\url{https://www.unece.org/fileadmin/DAM/stats/documents/ece/ces/ge.46/2013/Topic_1_ABS.pdf}
}
