% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ck_methods.R
\docType{methods}
\name{results}
\alias{results}
\alias{results,pert_table-method}
\title{Method to extract results}
\usage{
results(x, ...)

\S4method{results}{pert_table}(x, meanBeforeSum = TRUE)
}
\arguments{
\item{x}{input object}

\item{...}{additional parameters passed to methods}

\item{meanBeforeSum}{(logical) defines if for (continuous) variables the weighted means should be computed
before weighted sums or the other way round.}
}
\description{
Method to extract results
}
\seealso{
\url{https://www.unece.org/fileadmin/DAM/stats/documents/ece/ces/ge.46/2013/Topic_1_ABS.pdf}
}
