% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{testdata}
\alias{testdata}
\title{A real-world data set on household income and expenditures}
\format{
testdata: a data frame with 4580 observations on the following 15 variables.
\itemize{
\item \code{urbrur}: a numeric vector
\item \code{roof}: a numeric vector
\item \code{walls}: a numeric vector
\item \code{water}: a numeric vector
\item \code{electcon}: a numeric vector
\item \code{relat}: a numeric vector
\item \code{sex}: a numeric vector
\item \code{age}: a numeric vector
\item \code{hhcivil}: a numeric vector
\item \code{expend}: a numeric vector
\item \code{income}: a numeric vector
\item \code{savings}: a numeric vector
\item \code{ori_hid}: a numeric vector
\item \code{sampling_weight}: a numeric vector
\item \code{household_weights}: a numeric vector
}
}
\description{
4580 Obervations in 15 Variables; This dataset also contains sampling weights!
}
\examples{
data(testdata)
head(testdata)
}
\references{
The International Household Survey Network, www.ihsn.org
}
\keyword{datasets}
